// src/lib/enum.ts
export enum TRAITS {
  FIRST_TIME_TYPE = 1,
  DURATION_TYPE = 2,
  IS_PLAYING_TYPE = 3,
  IS_RETURNING_TYPE = 5,
  IS_MARATHON_TYPE = 6,
  NEW_RELEASE_DATE_TYPE = 7,
  STREAK_TYPE = 8,
  IS_TRENDING_TYPE = 9,
}

export enum CONTENT_TYPE {
  GAME = 1,
  MUSIC = 5,
  ACTIVITY = 8,
}

export enum APP_EXTRA {
  PLAYED_EXTRA = "played_game_extra",
  LISTENING_EXTRA = "listened_media_extra",
}